requirements.txt
opencv-python
Pillow
fpdf
numpy
yt-dlp